Fix download, use delivery to download apks
Make purchase request if app is not in user library. Deprecates purchase call
to download apps, use delivery only to get app download link.
Also added a check to show a warning in the terminal if using a wrong version code
to download the app and return with no response.